+++ src/app/components/login/login.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { AuthService } from 'src/app/services/auth.service';
+import { NgForm } from '@angular/forms';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-login',
@@ -7,7 +9,17 @@ import { AuthService } from 'src/app/services/auth.service';
   styleUrls: ['./login.component.scss'],
 })
 export class LoginComponent implements OnInit {
-  constructor(public authService: AuthService) {}
+  form = {
+    email: '',
+    password: '',
+  };
+
+  constructor(public authService: AuthService, private router: Router) {}
 
   ngOnInit(): void {}
+
+  async onSubmit() {
+    await this.authService.loginWithEmail(this.form.email, this.form.password);
+    this.router.navigate(['/customers']);
+  }
 }